refactor: 리뷰반영
- index 페이지 수동 매핑 컨트롤러로 동작하도록 수정
- 파일 개행
- 줄여써진 변수명 수정
